% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_and_label}
\alias{clean_and_label}
\title{Clean and Label Dataset}
\usage{
clean_and_label(dataset)
}
\arguments{
\item{dataset}{A data frame representing the Add Health dataset to be cleaned and labeled.}
}
\value{
A data frame with applied value labels.
}
\description{
This function applies labels from the built-in Add Health codebook to the given dataset.
}
