% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{replace_dont_know_refused}
\alias{replace_dont_know_refused}
\title{Replace "Don't know" or "Refused" with NA}
\usage{
replace_dont_know_refused(dataset)
}
\arguments{
\item{dataset}{A data frame representing the dataset where replacements should be made.}
}
\value{
A data frame with "Don't know" or "Refused" replaced by NA.
}
\description{
This function replaces all instances of "Don't know" or "Refused" in variables with NA.
}
